+++ src/core/include/antioch/chemical_mixture.h
@@ -37,8 +37,6 @@
 #include "antioch/chemical_species.h"
 #include "antioch/default_filename.h"
 #include "antioch/metaprogramming.h"
-#include "antioch/species_parsing.h"
-#include "antioch/ascii_parser.h"
 
 // C++
 #include <vector>
@@ -49,6 +47,10 @@
 
 namespace Antioch
 {
+  // Forward declarations
+  template<typename CoeffType>
+  class ParserBase;
+
   typedef unsigned int Species;
 
   //! Class storing chemical mixture properties
@@ -156,10 +158,10 @@ namespace Antioch
     template<typename VectorStateType>
     typename enable_if_c<
       has_size<VectorStateType>::value,
-      typename Antioch::value_type<VectorStateType>::type 
+      typename Antioch::value_type<VectorStateType>::type
     >::type
     R( const VectorStateType& mass_fractions ) const;
-    
+
     //! Molecular weight (molar mass) for species s in kg/mol
     CoeffType M( const unsigned int s ) const;
 
@@ -172,18 +174,18 @@ namespace Antioch
     template<typename VectorStateType>
     typename enable_if_c<
       has_size<VectorStateType>::value,
-      typename Antioch::value_type<VectorStateType>::type 
+      typename Antioch::value_type<VectorStateType>::type
     >::type
     M( const VectorStateType& mass_fractions ) const;
 
     //! Species mole fraction
-    /*! 
+    /*!
       Given mixture molar mass M and mass fraction for species,
       compute species mole fraction using the relationship
-      \f$ w_i = x_i \frac{M_i}{M} \f$ 
+      \f$ w_i = x_i \frac{M_i}{M} \f$
     */
     template<typename StateType>
-    ANTIOCH_AUTO(StateType) 
+    ANTIOCH_AUTO(StateType)
     X( const unsigned int species, const StateType& M,
        const StateType& mass_fraction ) const
     ANTIOCH_AUTOFUNC(StateType, mass_fraction*M/this->M(species))
@@ -195,11 +197,11 @@ namespace Antioch
     */
     template<typename VectorStateType>
     void X( typename Antioch::value_type<VectorStateType>::type M,
-	    const VectorStateType& mass_fractions, 
+	    const VectorStateType& mass_fractions,
 	    VectorStateType& mole_fractions ) const;
 
     //! Species molar density
-    /*! 
+    /*!
       Given total density rho and mass fraction for species,
       compute species moles per unit volume
     */
@@ -211,7 +213,7 @@ namespace Antioch
     ANTIOCH_AUTOFUNC(StateType, rho*mass_fraction/this->M(species))
 
     //! Species molar densities
-    /*! 
+    /*!
       Given total density rho and mass fractions for all species,
       compute moles per unit volume for all species
     */
@@ -226,7 +228,6 @@ namespace Antioch
     void build_inverse_name_map();
 
     std::vector<Species> _species_list;
-//    std::map<Species,unsigned int> _species_list_map;
     std::vector<ChemicalSpecies<CoeffType>*> _chemical_species;
     std::map<std::string,Species> _species_name_map;
     std::map<Species,std::string> _species_inv_name_map;
@@ -245,17 +246,10 @@ namespace Antioch
   template<typename CoeffType>
   inline
   const std::vector<Species>& ChemicalMixture<CoeffType>::species_list() const
-  { 
+  {
     return _species_list;
   }
 
-/*  template<typename CoeffType>
-  inline
-  const std::map<Species,unsigned int>& ChemicalMixture<CoeffType>::species_list_map() const
-  {
-    return _species_list_map;
-  }
-*/
   template<typename CoeffType>
   inline
   const std::vector<ChemicalSpecies<CoeffType>*>& ChemicalMixture<CoeffType>::chemical_species() const
@@ -321,151 +315,18 @@ namespace Antioch
   }
 
   template<typename CoeffType>
-  inline
-  ChemicalMixture<CoeffType>::ChemicalMixture(const std::string & filename, const bool verbose, 
-                                              const std::string & species_data,
-                                              const std::string & vibration_data,
-                                              const std::string & electronic_data)
-  {
-      ASCIIParser<CoeffType> parser(filename,verbose);
-
-      read_chemical_species_composition<CoeffType>(static_cast<ParserBase<CoeffType> *> (&parser), *this);
-
-      parser.change_file(species_data);
-      this->read_species_characteristics(&parser,species_data,vibration_data,electronic_data);
-
-      return;
-  }
-
-  template<typename CoeffType>
-  inline
-  ChemicalMixture<CoeffType>::ChemicalMixture(ParserBase<CoeffType> * parser,
-                                              const std::string & species_data,
-                                              const std::string & vibration_data,
-                                              const std::string & electronic_data)
-  {
-      read_chemical_species_composition<CoeffType>(parser, *this);
-
-      this->read_species_characteristics(parser,species_data,vibration_data,electronic_data);
-
-      return;
-  }
-
-  template<typename CoeffType>
-  inline
-  void ChemicalMixture<CoeffType>::read_species_characteristics(ParserBase<CoeffType> * parser,
-                                                                const std::string & /*species_data*/,
-                                                                const std::string & vibration_data,
-                                                                const std::string & electronic_data)
-  {
-   // species file is already in parser object
-     this->read_species_mandatory_characteristics(parser);
-
-    //... and any vibrational data
-     parser->change_file(vibration_data);
-     this->read_species_vibrational_characteristics(parser);
-
-    //... and any electronic data
-     parser->change_file(electronic_data);
-     this->read_species_electronic_characteristics(parser);
-
-    return;
-  }
-
-
-  template <typename CoeffType>
-  inline
-  void ChemicalMixture<CoeffType>::read_species_mandatory_characteristics(ParserBase<CoeffType> * parser)
-  {
-      read_species_data<CoeffType>(parser, *this);
-  }
-
-  template <typename CoeffType>
-  inline
-  void ChemicalMixture<CoeffType>::read_species_vibrational_characteristics(ParserBase<CoeffType> * parser)
-  {
-      read_species_vibrational_data<CoeffType>(parser, *this);
-  }
-
-  template <typename CoeffType>
-  inline
-  void ChemicalMixture<CoeffType>::read_species_electronic_characteristics(ParserBase<CoeffType> * parser)
-  {
-      read_species_electronic_data<CoeffType>(parser, *this);
-  }
-
-  template<typename CoeffType>
-  inline
-  ChemicalMixture<CoeffType>::ChemicalMixture(const std::vector<std::string>& species_list,
-                                              const bool verbose,
-                                              const std::string & species_data,
-                                              const std::string & vibration_data,
-                                              const std::string & electronic_data)
-  {
-      this->initialize_species(species_list);
-
-      Antioch::ASCIIParser<CoeffType> parser(species_data,verbose);
-
-      this->read_species_characteristics(static_cast<ParserBase<CoeffType> *>(&parser),species_data,vibration_data,electronic_data);
-
-      return;
-  }
-
-  template<typename CoeffType>
-  inline
-  void ChemicalMixture<CoeffType>::initialize_species( const std::vector<std::string>& species_list )
-  {
-    _chemical_species.resize( species_list.size(), NULL );
-    // Build up name map for all possible species
-    this->init_species_name_map(species_list);
-    
-    // Build up inverse name map
-    this->build_inverse_name_map();
-
-    // Populate species list for requested species
-    _species_list.reserve( species_list.size() );
-    for( unsigned int s = 0; s < species_list.size(); s++ )
-      {
-	if( _species_name_map.find( species_list[s] ) == _species_name_map.end() )
-	  {
-	    std::cerr << "Error in ChemicalMixture: Unknown species " << species_list[s] << std::endl;
-	    antioch_error();
-	  }
-
-	_species_list.push_back( _species_name_map.find( species_list[s] )->second );
-//	_species_list_map.insert( std::make_pair( _species_name_map.find( species_list[s] )->second, s ) );
-      }
-  }
-
-
-  template<typename CoeffType>
-  inline
-  ChemicalMixture<CoeffType>::~ChemicalMixture()
-  {
-    // Clean up all the ChemicalSpecies we stored
-    for( typename std::vector<ChemicalSpecies<CoeffType>* >::iterator it = _chemical_species.begin();
-	 it < _chemical_species.end(); ++it )
-      {
-	delete (*it);
-      }
-
-    return;
-  }
-
-
-  template<typename CoeffType>
   template<typename VectorStateType>
   inline
   typename enable_if_c<
     has_size<VectorStateType>::value,
-    typename Antioch::value_type<VectorStateType>::type 
+    typename Antioch::value_type<VectorStateType>::type
   >::type
   ChemicalMixture<CoeffType>::R( const VectorStateType& mass_fractions ) const
   {
     antioch_assert_equal_to( mass_fractions.size(), _chemical_species.size() );
     antioch_assert_greater( mass_fractions.size(), 0);
-    
-    typename Antioch::value_type<VectorStateType>::type 
+
+    typename Antioch::value_type<VectorStateType>::type
       R = mass_fractions[0]*this->R(0);
     for( unsigned int s = 1; s < mass_fractions.size(); s++ )
       {
@@ -481,13 +342,13 @@ namespace Antioch
   inline
   typename enable_if_c<
     has_size<VectorStateType>::value,
-    typename Antioch::value_type<VectorStateType>::type 
+    typename Antioch::value_type<VectorStateType>::type
   >::type
   ChemicalMixture<CoeffType>::M( const VectorStateType& mass_fractions ) const
   {
     antioch_assert_equal_to( mass_fractions.size(), _chemical_species.size() );
 
-    typename Antioch::value_type<VectorStateType>::type 
+    typename Antioch::value_type<VectorStateType>::type
       M = mass_fractions[0]/this->M(0);
     for( unsigned int s = 1; s < mass_fractions.size(); s++ )
       {
@@ -502,7 +363,7 @@ namespace Antioch
   template<typename VectorStateType>
   inline
   void ChemicalMixture<CoeffType>::X( typename Antioch::value_type<VectorStateType>::type M,
-				      const VectorStateType& mass_fractions, 
+				      const VectorStateType& mass_fractions,
 				      VectorStateType& mole_fractions ) const
   {
     antioch_assert_equal_to( mass_fractions.size(), _chemical_species.size() );
@@ -516,21 +377,6 @@ namespace Antioch
       {
 	mole_fractions[s] = this->X(s, M, mass_fractions[s]);
       }
-
-    return;
-  }
-
-
-  template<typename CoeffType>
-  inline
-  void ChemicalMixture<CoeffType>::add_species( const unsigned int index,
-					        const std::string& name,
-					        CoeffType mol_wght,
-					        CoeffType h_form,
-					        CoeffType n_tr_dofs, int charge)
-  {
-    _chemical_species[index] =
-      new ChemicalSpecies<CoeffType>(name, mol_wght, h_form, n_tr_dofs, charge);
 
     return;
   }
@@ -551,30 +397,6 @@ namespace Antioch
                                                                 const unsigned int ndg_e )
   {
     (_chemical_species[index])->add_electronic_data(theta_e, ndg_e);
-  }
-  
-  template<typename CoeffType>
-  inline
-  void ChemicalMixture<CoeffType>::init_species_name_map(const std::vector<std::string> & species_list)
-  {
-    _species_name_map.clear();
-    for(unsigned int s = 0; s < species_list.size(); s++)
-    {
-        _species_name_map[species_list[s]] = s;
-    }
-  }
-
-  template<typename CoeffType>
-  inline
-  void ChemicalMixture<CoeffType>::build_inverse_name_map()
-  {
-    for( std::map<std::string,Species>::const_iterator it = _species_name_map.begin();
-	 it != _species_name_map.end(); ++it )
-      {
-	_species_inv_name_map.insert( std::make_pair( it->second, it->first ) );
-      }
-
-    return;
   }
 
   template<typename CoeffType>
