+++ src/parsing/include/antioch/nasa_mixture_parsing.h
@@ -24,23 +24,14 @@
 //-----------------------------------------------------------------------el-
 //--------------------------------------------------------------------------
 
-#ifndef ANTIOCH_NASA_MIXTURE_ASCII_PARSING_H
-#define ANTIOCH_NASA_MIXTURE_ASCII_PARSING_H
-
-// C++
-#include <iostream>
-#include <fstream>
-#include <vector>
-#include <string>
-#include <sstream>
+#ifndef ANTIOCH_NASA_MIXTURE_PARSING_H
+#define ANTIOCH_NASA_MIXTURE_PARSING_H
 
 // Antioch
-#include "antioch/input_utils.h"
-#include "antioch/chemical_mixture.h"
 #include "antioch/parsing_enum.h"
-#include "antioch/ascii_parser.h"
-#include "antioch/xml_parser.h"
-#include "antioch/chemkin_parser.h"
+
+// C++
+#include <string>
 
 namespace Antioch
 {
@@ -50,53 +41,19 @@ namespace Antioch
 
   // definition problems du to backward compatibilities...
   /*
-     ASCIIParser -> CEAThermodynamics 
-                 ->  "antioch/cea_mixture_ascii_parsing.h" 
-                 ->  "antioch/cea_mixture_parsing.h" 
-                 ->  "antioch/nasa_mixture_parsing.h" 
+     ASCIIParser -> CEAThermodynamics
+                 ->  "antioch/cea_mixture_ascii_parsing.h"
+                 ->  "antioch/cea_mixture_parsing.h"
+                 ->  "antioch/nasa_mixture_parsing.h"
                  -> ASCIIParser
    */
-  template <class NumericType>
-  class ASCIIParser;
 
   template<class NumericType, typename CurveType >
-  void read_nasa_mixture_data( NASAThermoMixture<NumericType, CurveType > & thermo, const std::string &filename, ParsingType = ASCII, bool verbose = true );
-
- 
-  /* ------------------------- Inline Functions -------------------------*/
-  template<class NumericType, typename CurveType>
-  inline
-  void read_nasa_mixture_data( NASAThermoMixture<NumericType, CurveType >& thermo, const std::string &filename, ParsingType type, bool verbose )
-  {
-    
-    ParserBase<NumericType> * parser(NULL);
-    switch(type)
-    {
-      case ASCII:
-         parser = new ASCIIParser<NumericType>(filename,verbose);
-         break;
-      case CHEMKIN:
-         parser = new ChemKinParser<NumericType>(filename,verbose);
-         break;
-      case XML:
-         parser = new XMLParser<NumericType>(filename,verbose);
-         break;
-      default:
-         antioch_parsing_error("unknown type");
-    }
-
-   parser->read_thermodynamic_data(thermo);
-
-   // Make sure we actually populated everything
-   if( !thermo.check() )
-     {
-	std::cerr << "Error: NASA table not fully populated" << std::endl;
-	antioch_error();
-      }
-
-    return;
-  }
+  void read_nasa_mixture_data( NASAThermoMixture<NumericType, CurveType >& thermo,
+                               const std::string &filename,
+                               ParsingType = ASCII,
+                               bool verbose = true );
 
 } // end namespace Antioch
 
-#endif // ANTIOCH_NASA_MIXTURE_ASCII_PARSING_H
+#endif // ANTIOCH_NASA_MIXTURE_PARSING_H
