+++ src/transport/include/antioch/transport_mixture.h
@@ -34,29 +34,16 @@
 // Antioch
 #include "antioch/antioch_asserts.h"
 #include "antioch/transport_species.h"
-#include "antioch/metaprogramming.h"
 #include "antioch/chemical_mixture.h"
 #include "antioch/parsing_enum.h"
-#include "antioch/transport_species_parsing.h"
-#include "antioch/default_filename.h"
-#include "antioch/parser_base.h"
 
 // C++
 #include <vector>
 #include <string>
+#include <map>
 
 namespace Antioch
 {
-
-  template <typename CoeffType>
-  class ASCIIParser;
-
-  template <typename CoeffType>
-  class ChemKinParser;
-
-  template <typename CoeffType>
-  class XMLParser;
-
   //! Class storing chemical mixture properties
   /*!
     This class manages the list of TransportSpecies for a requested set
@@ -104,7 +91,7 @@ namespace Antioch
 
   protected:
 
-    const ChemicalMixture<CoeffType>        & _chemical_mixture;
+    const ChemicalMixture<CoeffType>& _chemical_mixture;
 
     std::vector<TransportSpecies<CoeffType>*> _transport_species;
 
@@ -166,108 +153,6 @@ namespace Antioch
     antioch_assert(_transport_species[s]);
 
     return (*_transport_species[s]);
-  }
-
-  template<typename CoeffType>
-  inline
-  TransportMixture<CoeffType>::TransportMixture( const ChemicalMixture<CoeffType>& chem_mix,
-                                                                 const std::string & filename, bool verbose, ParsingType type )
-    : _chemical_mixture( chem_mix),
-      _transport_species(_chemical_mixture.n_species(), NULL )
-  {
-
-   ParserBase<CoeffType> * parser(NULL);
-    switch(type)
-    {
-      case ASCII:
-         parser = new ASCIIParser<CoeffType>(filename,verbose);
-         break;
-      case CHEMKIN:
-         parser = new ChemKinParser<CoeffType>(filename,verbose);
-         break;
-      case XML:
-         parser = new XMLParser<CoeffType>(filename,verbose);
-         break;
-      default:
-         antioch_parsing_error("unknown type");
-    }
-
-    // Now read in transport properties for the requested species and stash
-    read_transport_species_data(parser,*this);
-
-    // check we have everyone requested
-    for( unsigned int s = 0; s < _transport_species.size(); ++s )
-      {
-        if(!_transport_species[s]) // it is not mandatory, Blottner or Sutherland are self-sufficient
-        {
-           std::cerr << "Warning: missing transport data for species " << _chemical_mixture.species_inverse_name_map().at(
-                                                                          _chemical_mixture.species_list()[s]) << "\n"
-                     << "Be sure to use a transport model that does not require the default data"
-                     << std::endl;
-        }
-      }
-
-    delete parser;
-
-    return;
-
-  }
-
-  template<typename CoeffType>
-  inline
-  TransportMixture<CoeffType>::TransportMixture( const ChemicalMixture<CoeffType> &mixture,
-                                                                 ParserBase<CoeffType> * parser)
-    : _chemical_mixture( mixture),
-      _transport_species(_chemical_mixture.n_species(), NULL )
-  {
-    // Now read in transport properties for the requested species and stash
-    read_transport_species_data(parser,*this);
-
-    // check we have everyone requested
-    for( unsigned int s = 0; s < _transport_species.size(); ++s )
-      {
-        if(!_transport_species[s]) // it is not mandatory, Blottner or Sutherland are self-sufficient
-        {
-           std::cerr << "Warning: missing transport data for species " << _chemical_mixture.species_inverse_name_map().at(
-                                                                          _chemical_mixture.species_list()[s]) << "\n"
-                     << "Be sure to use a transport model that does not require the default data"
-                     << std::endl;
-        }
-      }
-
-    return;
-
-  }
-
-  template<typename CoeffType>
-  inline
-  TransportMixture<CoeffType>::~TransportMixture()
-  {
-    // Clean up all the TransportSpecies we stored
-    for( typename std::vector<TransportSpecies<CoeffType>* >::iterator it = _transport_species.begin();
-         it < _transport_species.end(); ++it )
-      {
-        delete (*it);
-      }
-
-    return;
-  }
-
-  template<typename CoeffType>
-  inline
-  void TransportMixture<CoeffType>::add_species( const unsigned int index,
-                                                 CoeffType LJ_depth,
-                                                 CoeffType LJ_diameter,
-                                                 CoeffType dipole_moment,
-                                                 CoeffType polarizability,
-                                                 CoeffType rotational_relaxation,
-                                                 CoeffType mass)
-  {
-    Species name_enum = _chemical_mixture.species_list()[index];
-    _transport_species[index] =
-      new TransportSpecies<CoeffType>(name_enum, LJ_depth, LJ_diameter, dipole_moment, polarizability, rotational_relaxation,mass);
-
-    return;
   }
 
 } // end namespace Antioch
